add express and supertest for integration tests
