+++ package.json
@@ -26,12 +26,14 @@
     "deep-freeze": "0.0.1",
     "eslint": "4.4.1",
     "eslint-plugin-react": "7.2.1",
+    "express": "4.16.2",
     "pre-commit": "1.2.2",
     "prop-types": "15.5.10",
     "ramda": "0.24.1",
     "react": "15.6.1",
     "react-dom": "15.6.1",
     "riteway": "2.0.3",
+    "supertest": "3.0.0",
     "tape": "4.6.3",
     "watch": "1.0.2"
   },
